requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.8
idna==3.6
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
stripe==7.9.0
typing-extensions==4.9.0
urllib3==2.1.0
whitenoise==6.6.0
